Move ActivityList to components
