+++ src/components/ActivityList/styles.js
